+++ src/main/java/legend/game/inventory/screens/EquipmentScreen.java
@@ -201,7 +201,7 @@ public class EquipmentScreen extends MenuScreen {
   }
 
   @Override
-  protected InputPropagation mouseMove(final int x, final int y) {
+  protected InputPropagation mouseMove(final double x, final double y) {
     if(super.mouseMove(x, y) == InputPropagation.HANDLED) {
       return InputPropagation.HANDLED;
     }
@@ -211,7 +211,7 @@ public class EquipmentScreen extends MenuScreen {
     }
 
     for(int slot = 0; slot < Math.min(4, this.equipmentCount - this.slotScroll); slot++) {
-      if(this.selectedSlot != slot && MathHelper.inBox(x, y, 212, this.menuHighlightPositionY(slot), 139, 15)) {
+      if(this.selectedSlot != slot && MathHelper.inBox((int)x, (int)y, 212, this.menuHighlightPositionY(slot), 139, 15)) {
         playMenuSound(1);
         this.selectedSlot = slot;
         this.itemHighlight.y_44 = this.menuHighlightPositionY(slot);
@@ -223,7 +223,7 @@ public class EquipmentScreen extends MenuScreen {
   }
 
   @Override
-  protected InputPropagation mouseClick(final int x, final int y, final int button, final Set<InputMod> mods) {
+  protected InputPropagation mouseClick(final double x, final double y, final int button, final Set<InputMod> mods) {
     if(super.mouseClick(x, y, button, mods) == InputPropagation.HANDLED) {
       return InputPropagation.HANDLED;
     }
@@ -233,7 +233,7 @@ public class EquipmentScreen extends MenuScreen {
     }
 
     for(int slot = 0; slot < Math.min(4, this.equipmentCount - this.slotScroll); slot++) {
-      if(MathHelper.inBox(x, y, 212, this.menuHighlightPositionY(slot), 139, 15)) {
+      if(MathHelper.inBox((int)x, (int)y, 212, this.menuHighlightPositionY(slot), 139, 15)) {
         this.selectedSlot = slot;
         this.itemHighlight.y_44 = this.menuHighlightPositionY(slot);
